requirements.txt
bigquery==0.0.38
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
dacktool==0.0.7
dbstream==0.1.21
google-api-core==2.14.0
google-api-python-client==1.7.11
google-auth==2.24.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.23.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.7.2
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-oauth2-tool==0.0.3
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
googleauthentication==0.0.17
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
idna==3.6
ijson==3.2.3
libcst==1.1.0
loguru==0.7.2
mypy-extensions==1.0.0
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pandas==1.3.4
proto-plus==1.22.3
protobuf==4.25.1
py4j==0.10.9.7
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==2.1.0
